review inna portal, but at what cost
