+++ client/containers/Pub/PubHeader/Review/review.scss
@@ -1,6 +1,5 @@
 .review-component {
 	padding-bottom: 50px;
-	// width: fit-content;
 
 	.review-border {
 		border-style: solid;
