+++ client/containers/Pub/PubHeader/reviewHeaderSticky.scss
@@ -1,5 +1,5 @@
 @import 'styles/variables.scss';
-@import '../../pub.scss';
+@import '../pub.scss';
 
 .review-header-sticky-component {
 	position: absolute;
